chore: moving nft test fixtures to their own file
