+++ src/nft/components/profile/view/ViewMyNftsAsset.test.tsx
@@ -1,4 +1,4 @@
-import { TEST_NFT_WALLET_ASSET } from 'test-utils/constants'
+import { TEST_NFT_WALLET_ASSET } from 'test-utils/nftFixtures'
 import { render } from 'test-utils/render'
 
 import { ViewMyNftsAsset } from './ViewMyNftsAsset'
